# PacketProcessing/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Kanat Packet Processing API",
    "version": "v1",
    "description": "API for real-time packet capture and playback analysis",
    "contact": {
      "name": "Kanat Development Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:10901",
      "description": "Development server"
    }
  ],
...
  "paths": {
    "/api/range/mode/{mode}": {
      "put": {
        "tags": ["Range Management"],
        "summary": "Change application mode",
        "description": "Changes the application mode between Realtime and Playback with validation",
        "operationId": "ChangeMode",
        "parameters": [
          {
            "name": "mode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["Realtime", "Playback"]
            },
            "description": "The target mode to switch to"
          }
        ],
        "responses": {
          "200": {
            "description": "Mode changed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult"
                },
                "example": {
                  "success": true,
                  "data": null,
                  "errorMessage": null,
                  "statusCode": 200,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Invalid mode transition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Cannot switch to Playback mode while capture may be active. Please ensure capture is stopped.",
                  "statusCode": 400,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/range/mode": {
      "get": {
        "tags": ["Range Management"],
        "summary": "Get current mode",
        "description": "Gets the current application mode",
        "operationId": "GetMode",
        "responses": {
          "200": {
            "description": "Current mode retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultString"
                },
                "example": {
                  "success": true,
                  "data": "Realtime",
                  "errorMessage": null,
                  "statusCode": 200,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultString"
                }
              }
            }
          }
        }
      }
    },
    "/api/range/dev/realtime/start/{deviceName}": {
      "post": {
        "tags": ["Development"],
        "summary": "Start realtime services",
        "description": "Starts all realtime capture services for the specified device",
        "operationId": "StartAllServices",
        "parameters": [
          {
            "name": "deviceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Name of the network device to start capture on",
            "example": "eth0"
          }
        ],
        "responses": {
          "200": {
            "description": "Services started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult"
                }
              }
            }
          },
          "500": {
            "description": "Failed to start services",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Failed to start services",
                  "statusCode": 500,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          }
        }
      }
    },
    "/api/range/dev/realtime/stop": {
      "delete": {
        "tags": ["Development"],
        "summary": "Stop realtime services (Development)",
        "description": "Stops all realtime capture services (development-only endpoint)",
        "operationId": "DevStopAllServices",
        "responses": {
          "200": {
            "description": "Services stopped successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult"
                }
              }
            }
          },
          "500": {
            "description": "Failed to stop services",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Failed to stop services",
                  "statusCode": 500,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          }
        }
      }
    },
    "/api/range/realtime/start": {
      "post": {
        "tags": ["Range Management"],
        "summary": "Start realtime with configuration",
        "description": "Starts realtime capture using a full Range configuration (device and endpoints).",
        "operationId": "StartRealtimeWithConfig",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RangeStartRequest" },
              "example": {
                "description": "Range session",
                "config": {
                  "bpfConfig": {
                    "device": "any",
                    "motion": [ { "ip": "132.8.7.125", "port": 5000 } ],
                    "safety": [ { "ip": "132.8.7.101", "port": 5000 }, { "ip": "132.8.7.102", "port": 5001 } ],
                    "onVIF": [ { "ip": "132.8.7.121", "port": 5000 } ]
                  },
                  "mtxConfig": { "ip": "127.0.0.1", "port": 8554 },
                  "cams": [ { "alias": "Default", "isRecording": false } ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Realtime started",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultRange" } } }
          },
          "500": {
            "description": "Failed to start realtime",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResponseResultRange" } } }
          }
        }
      }
    },
    "/api/range/realtime/stop": {
      "delete": {
        "tags": ["Range Management"],
        "summary": "Stop realtime services",
        "description": "Stops all realtime capture services",
        "operationId": "StopAllServices",
        "responses": {
          "200": {
            "description": "Services stopped successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultRange"
                }
              }
            }
          },
          "500": {
            "description": "Failed to stop services",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultRange"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Failed to stop services",
                  "statusCode": 500,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          }
        }
      }
    },
    "/api/range/playback/pace/{pace}": {
      "put": {
        "tags": ["Range Management"],
        "summary": "Set playback pace",
        "description": "Sets the playback pace (speed multiplier)",
        "operationId": "SetPlaybackPace",
        "parameters": [
          {
            "name": "pace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double",
              "minimum": 0.1
            },
            "description": "Playback pace multiplier (e.g., 1.0 = normal speed, 2.0 = double speed)",
            "example": 1.5
          }
        ],
        "responses": {
          "200": {
            "description": "Playback pace set successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pace value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Pace must be greater than 0",
                  "statusCode": 400,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          },
          "500": {
            "description": "Failed to set playback pace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/range/realtime/devices": {
      "get": {
        "tags": ["Range Status"],
        "summary": "Get available devices",
        "description": "Gets the list of available network devices for packet capture",
        "operationId": "GetAvailableDevices",
        "responses": {
          "200": {
            "description": "Available devices retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultStringArray"
                },
                "example": {
                  "success": true,
                  "data": ["eth0", "wlan0", "lo"],
                  "errorMessage": null,
                  "statusCode": 200,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get available devices",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResultStringArray"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Failed to get available devices",
                  "statusCode": 500,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          }
        }
      }
    },
    "/api/range/reset": {
      "post": {
        "tags": ["Range Status"],
        "summary": "Reset statistics",
        "description": "Resets statistics for the current mode",
        "operationId": "ResetStatistics",
        "responses": {
          "200": {
            "description": "Statistics reset successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult"
                }
              }
            }
          },
          "500": {
            "description": "Failed to reset statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult"
                },
                "example": {
                  "success": false,
                  "data": null,
                  "errorMessage": "Failed to reset statistics",
                  "statusCode": 500,
                  "timestamp": "2024-01-15T10:30:00Z"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "EndpointSpecification": {
        "type": "object",
        "properties": {
          "ip": { "type": "string", "nullable": true },
          "port": { "type": "integer", "nullable": true }
        }
      },
      "BPFConfDto": {
        "type": "object",
        "properties": {
          "safety": { "type": "array", "items": { "$ref": "#/components/schemas/EndpointSpecification" } },
          "motion": { "type": "array", "items": { "$ref": "#/components/schemas/EndpointSpecification" } },
          "onVIF":  { "type": "array", "items": { "$ref": "#/components/schemas/EndpointSpecification" } },
          "device": { "type": "string" }
        },
        "required": ["device"]
      },
      "CameraConfDto": {
        "type": "object",
        "properties": {
          "alias": { "type": "string" },
          "isRecording": { "type": "boolean" }
        },
        "required": ["alias", "isRecording"]
      },
      "RangeConfig": {
        "type": "object",
        "properties": {
          "bpfConfig": { "$ref": "#/components/schemas/BPFConfDto" },
          "mtxConfig": { "$ref": "#/components/schemas/EndpointSpecification" },
          "cams": { "type": "array", "items": { "$ref": "#/components/schemas/CameraConfDto" } }
        },
        "required": ["bpfConfig"]
      },
      "RangeStartRequest": {
        "type": "object",
        "properties": {
          "description": { "type": "string" },
          "config": { "$ref": "#/components/schemas/RangeConfig" }
        },
        "required": ["config"]
      },
      "ResponseResult": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates whether the operation was successful"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "description": "The data payload (null if operation failed)"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true,
            "description": "Error message (null if operation succeeded)"
          },
          "statusCode": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of the response"
          }
        },
        "required": ["success", "statusCode", "timestamp"]
      },
      "ResponseResultString": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates whether the operation was successful"
          },
          "data": {
            "type": "string",
            "nullable": true,
            "description": "The string data payload"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true,
            "description": "Error message (null if operation succeeded)"
          },
          "statusCode": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of the response"
          }
        },
        "required": ["success", "statusCode", "timestamp"]
      },
      "ResponseResultStringArray": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates whether the operation was successful"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "description": "The array of strings data payload"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true,
            "description": "Error message (null if operation succeeded)"
          },
          "statusCode": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of the response"
          }
        },
        "required": ["success", "statusCode", "timestamp"]
      },
...
      "RangeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the range"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the range was created"
          },
          "start": {
            "type": "integer",
            "format": "int64",
            "description": "Start position or time of the range"
          },
          "end": {
            "type": "integer",
            "format": "int64",
            "description": "End position or time of the range"
          },
          "description": {
            "type": "string",
            "description": "Description of the range"
          }
        },
        "required": ["id", "timestamp", "start", "end", "description"]
      },
      "ResponseResultRange": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates whether the operation was successful"
          },
          "data": {
            "$ref": "#/components/schemas/RangeDto",
            "nullable": true,
            "description": "The range data payload"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true,
            "description": "Error message (null if operation succeeded)"
          },
          "statusCode": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of the response"
          }
        },
        "required": ["success", "statusCode", "timestamp"]
      },
...
    }
  },
...
}